6.1. Rename withAudioPlayer HOC to withActivePlayer for the sake of consistency
